<==== Mercurial data ====>
Branch: feature_3948_webclient

+++ common/src/utils/media/media_stream_cache.cpp
@@ -379,7 +379,7 @@ void MediaStreamCache::unblockData( int blockingID )
     m_dataBlockings.erase( it );
 }
 
-size_t MediaStreamCache::inactivityPeriod() const
+qint64 MediaStreamCache::inactivityPeriod() const
 {
     QMutexLocker lk( &m_mutex );
     return m_inactivityTimer.elapsed();
